refactor: split memory/schema.ts into domain-specific schema files
