+++ assistant/src/memory/schema/index.ts
@@ -1,0 +1,8 @@
+export * from "./calls.js";
+export * from "./contacts.js";
+export * from "./conversations.js";
+export * from "./guardian.js";
+export * from "./infrastructure.js";
+export * from "./memory-core.js";
+export * from "./notifications.js";
+export * from "./tasks.js";
